Move some methods to the utils package
